refactor(client): implement barrel exports for cleaner imports
- Create index.ts files for components, hooks, stores, and lib
- Update imports across the application to use barrel files
- Change BannerAd to named export for consistency
