+++ client/src/app/cms/page.tsx
@@ -1,4 +1,4 @@
-import { CMSDashboard } from "@/components/cms/CMSDashboard";
+import { CMSDashboard } from "@/components/cms";
 
 export default function AdminPage() {
   return <CMSDashboard />;
