+++ client/src/components/ads/index.ts
@@ -1,0 +1,2 @@
+export * from "./AdSlot";
+export * from "./BannerAd";
