+++ client/src/components/pages/index.ts
@@ -1,0 +1,7 @@
+export * from "./EscrowPage";
+export * from "./LandingPage";
+export * from "./LoginPage";
+export * from "./SendPage";
+export * from "./SwapPage";
+export * from "./TrendViewPage";
+export * from "./WalletDashboard";
