+++ client/src/components/ui/index.ts
@@ -1,0 +1,8 @@
+export * from "./button";
+export * from "./card";
+export * from "./dialog";
+export * from "./input";
+export * from "./label";
+export * from "./sonner";
+export * from "./tabs";
+export * from "./textarea";
