+++ client/src/components/ui/label.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as LabelPrimitive from "@radix-ui/react-label";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib";
 
 function Label({
     className,
